# src/main/resources/spring-frontend/swagger.yaml
...
swagger: "2.0"
info:
    description: This is a simple API
    version: 1.0.0
    title: Books Api Server
    contact:
        email: mail@lunpully.com
    license:
        name: Apache 2.0
        url: http://www.apache.org/licenses/LICENSE-2.0.html
host: virtserver.swaggerhub.com
basePath: /lunpully/bookshop/1.0.0
schemes:
    - https
    - http
consumes:
    - application/json
produces:
    - application/json
paths:
    /books/recommended:
        get:
            tags:
                - recomended
            summary: Show recommended books
            operationId: recomended
            parameters:
                -   name: offset
                    in: query
                    description: The number of items to skip before starting to collect the result set.
                    required: false
                    type: integer
                -   name: limit
                    in: query
                    description: The numbers of items to return.
                    required: false
                    type: integer
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/Data'
                "400":
                    description: Error
                    schema:
                        $ref: '#/definitions/Error'
    /books/recent:
        get:
            tags:
                - recent
            summary: Show recent books
            operationId: recent
            parameters:
                -   name: from
                    in: query
                    required: false
                    type: string
                    format: date-time
                -   name: to
                    in: query
                    required: false
                    type: string
                    format: date-time
                -   name: offset
                    in: query
                    description: The number of items to skip before starting to collect the result set.
                    required: false
                    type: integer
                -   name: limit
                    in: query
                    description: The numbers of items to return.
                    required: false
                    type: integer
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/Data'
                "400":
                    description: Error
                    schema:
                        $ref: '#/definitions/Error'
...
    /search/{query}:
        get:
            tags:
                - search
            summary: Search
            operationId: search
            parameters:
                -   name: query
                    in: path
                    required: true
                    type: string
                -   name: offset
                    in: query
                    description: The number of items to skip before starting to collect the result set.
                    required: false
                    type: integer
                -   name: limit
                    in: query
                    description: The numbers of items to return.
                    required: false
                    type: integer
            responses:
                "200":
                    description: Successful response
                    schema:
                        $ref: '#/definitions/Data'
                "400":
                    description: Error
                    schema:
                        $ref: '#/definitions/Error'
...
definitions:
...
    Data:
        type: object
        properties:
            count:
                type: integer
                format: int
                example: 20
            books:
                $ref: '#/definitions/Book'
#                type: array
#                items:
#                    $ref: '#/definitions/BookEntity'
    Book:
        type: object
        properties:
            id:
                type: integer
                format: int
            slug:
                type: string
            image:
                type: string
            authors:
                type: string
            title:
                type: string
            discount:
                type: integer
                format: int
            isBestseller:
                type: boolean
            rating:
                type: integer
                format: int
            status:
                type: string
            price:
                type: integer
                format: int
            discountPrice:
                type: integer
                format: int
        example:
            -   id: 1
                slug: 759144
                title: 'Джой Уильямс: Подменыш'
                image: /assets/img/content/main/card.jpg
                authors: Уильямс Джой
                discount: 18
                isBestseller: true
                rating: 5
                status: KEPT
                price: 624
                discountPrice: 512
            -   id: 2
                slug: 764415
                title: Python и наука о данных для чайников
                authors: Диалектика
                image: /assets/img/content/main/card.jpg
                discount: 18
                isBestseller: false
                rating: 5
                status: PAID
                price: 1921
                discountPrice: 1575
            -   id: 3
                slug: 764416
                title: Тревожные расстройства и фобии. Когнитивный подход
                authors:
                image: /assets/img/content/main/card.jpg
                discount: 18
                isBestseller: false
                rating: 5
                status: false
                price: 3200
                discountPrice: 2624
            -   id: 4
                slug:  764416
                title: 'Донни Кейтс: Серебряный Сёрфер. Тьма'
                authors: Кейт Донни
                image: /assets/img/content/main/card.jpg
                discount: 18
                isBestseller: false
                rating: 5
                status: CART
                price: 978
                discountPrice: 802
            -   id: 5
                slug: 764416
                title: 'Rudyard Kipling: The Jungle (+CD)'
                authors: Киплинг Редьярд Джозеф
                image: /assets/img/content/main/card.jpg
                discount: 40
                isBestseller: false
                rating: 5
                status: false
                price: 429
                discountPrice: 257
            -   id: 6
                slug: 764416
                title: "L. Alexander: K's First Case"
                authors: Alexander L. G.
                image: /assets/img/content/main/card.jpg
                discount: 40
                isBestseller: false
                rating: 5
                status: KEPT
                price: 329
                discountPrice: 197
    
            -   id: 7
                slug: 764416
                title: "Bernard Smith: The Last Photo"
                authors: Smith Bernard
                image: /assets/img/content/main/card.jpg
                discount: 40
                isBestseller: true
                rating: 5
                status: false
                price: 329
                discountPrice: 197
            -   id: 8
                slug: 759144
                title: 'Джой Уильямс: Подменыш'
                image: /assets/img/content/main/card.jpg
                authors: Уильямс Джой
                discount: 18
                isBestseller: true
                rating: 5
                status: KEPT
                price: 624
                discountPrice: 512
            -   id: 9
                slug: 764415
                title: Python и наука о данных для чайников
                authors: Диалектика
                image: /assets/img/content/main/card.jpg
                discount: 18
                isBestseller: false
                rating: 5
                status: PAID
                price: 1921
                discountPrice: 1575
            -   id: 10
                slug: 764416
                title: Тревожные расстройства и фобии. Когнитивный подход
                authors:
                image: /assets/img/content/main/card.jpg
                discount: 18
                isBestseller: false
                rating: 5
                status: false
                price: 3200
                discountPrice: 2624
            -   id: 11
                slug:  764416
                title: 'Донни Кейтс: Серебряный Сёрфер. Тьма'
                authors: Кейт Донни
                image: /assets/img/content/main/card.jpg
                discount: 18
                isBestseller: false
                rating: 5
                status: CART
                price: 978
                discountPrice: 802
            -   id: 12
                slug: 764416
                title: 'Rudyard Kipling: The Jungle (+CD)'
                authors: Киплинг Редьярд Джозеф
                image: /assets/img/content/main/card.jpg
                discount: 40
                isBestseller: false
                rating: 5
                status: false
                price: 429
                discountPrice: 257
            -   id: 13
                slug: 764416
                title: "L. Alexander: K's First Case"
                authors: Alexander L. G.
                image: /assets/img/content/main/card.jpg
                discount: 40
                isBestseller: false
                rating: 5
                status: KEPT
                price: 329
                discountPrice: 197
    
            -   id: 14
                slug: 764416
                title: "Bernard Smith: The Last Photo"
                authors: Smith Bernard
                image: /assets/img/content/main/card.jpg
                discount: 40
                isBestseller: true
                rating: 5
                status: false
                price: 329
                discountPrice: 197
            -   id: 15
                slug: 759144
                title: 'Джой Уильямс: Подменыш'
                image: /assets/img/content/main/card.jpg
                authors: Уильямс Джой
                discount: 18
                isBestseller: true
                rating: 5
                status: KEPT
                price: 624
                discountPrice: 512
            -   id: 16
                slug: 764415
                title: Python и наука о данных для чайников
                authors: Диалектика
                image: /assets/img/content/main/card.jpg
                discount: 18
                isBestseller: false
                rating: 5
                status: PAID
                price: 1921
                discountPrice: 1575
            -   id: 17
                slug: 764416
                title: Тревожные расстройства и фобии. Когнитивный подход
                authors:
                image: /assets/img/content/main/card.jpg
                discount: 18
                isBestseller: false
                rating: 5
                status: false
                price: 3200
                discountPrice: 2624
            -   id: 18
                slug:  764416
                title: 'Донни Кейтс: Серебряный Сёрфер. Тьма'
                authors: Кейт Донни
                image: /assets/img/content/main/card.jpg
                discount: 18
                isBestseller: false
                rating: 5
                status: CART
                price: 978
                discountPrice: 802
            -   id: 19
                slug: 764416
                title: 'Rudyard Kipling: The Jungle (+CD)'
                authors: Киплинг Редьярд Джозеф
                image: /assets/img/content/main/card.jpg
                discount: 40
                isBestseller: false
                rating: 5
                status: false
                price: 429
                discountPrice: 257
            -   id: 20
                slug: 764416
                title: "L. Alexander: K's First Case"
                authors: Alexander L. G.
                image: /assets/img/content/main/card.jpg
                discount: 40
                isBestseller: false
                rating: 5
                status: KEPT
                price: 329
                discountPrice: 197
...
    Error:
        type: object
        properties:
            result:
                type: boolean
                example: true
            error:
                type: string
                example: Какая-то ошибка
...